+++ src/app/(basic)/_component/Footer/Footer.tsx
