+++ src/app/(basic)/_component/Header/Header.component.spec.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@/src/component/Header/Header'
+import { Header } from '@/src/app/(basic)/_component/Header/Header'
 import { render, screen } from '@testing-library/react'
 import { describe, test, expect } from 'vitest'
